Add delimiter parameter to cudf::strings::capitalize() (#8620)
Closes #8597 

Add a parameter to the `cudf::strings::capitalize()` function to support capitalizing characters after a specified delimiter. This should meet the requirements of #8597 by passing a single ' ' character string as follows:

```
    auto results = cudf::strings::capitalize(strings_view, std::string(" "));
```

The new parameter has a default of empty string that keeps the current behavior so no updates are required to python/cython layer. The new parameter requires this PR to be a breaking change.

The source code for `title()` and `capitalize()` were further refactored using the CRTP pattern since much of the main internal code logic is the same.

This PR also includes additional gtests for this new parameter as well as adding some missing tests for empty columns.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/8620