Support header with embedded delimiter in csv writer (#8798)
Closes #8713 

If the column delimiter (e.g. comma) appears in a column header, then the header string must be enclosed in double-quotes when written to the output csv file. The logic that writes the header line is updated in this PR to handle an embedded delimiter, terminator, or double-quote character. Also added a gtest with a comma inside a header string.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Conor Hoekstra (https://github.com/codereport)
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/cudf/pull/8798