Remove _source_index from MultiIndex (#9191)
This PR is the first in a sequence refactoring the `MultiIndex` class, which is currently in a halfway state of making use of some facilities common to other Frames without using them consistently. This PR removes the internal `_source_data` parameter, which creates a thin `DataFrame` wrapper around a `MultiIndex` to provide access to those methods. Since `MultiIndex` is now clearly defined as a `Frame`, this PR moves us towards a much simpler and more uniform design where we use inherited `Frame` methods directly. This approach is significantly easier to maintain than treating a `MultiIndex` as being "backed" by a `DataFrame`. This approach also comes with various performance benefits. While relatively cheap, the internal conversion of a `MultiIndex` to a `DataFrame` is not free, and in cases where we need to return a modified `MultiIndex` the conversion _back_ from a `DataFrame` is significantly more complex. Moreover, in a number of cases we were using less efficient methods than those we could use directly on the `MultiIndex`.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - H. Thomson Comer (https://github.com/thomcom)
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Sheilah Kirui (https://github.com/skirui-source)

URL: https://github.com/rapidsai/cudf/pull/9191