Update Java nvcomp JNI bindings to nvcomp 2.x API (#9384)
This PR Closes #8336 

The Java nvcomp JNI bindings are updated to the nvcomp 2.x API and the nvcomp library is no longer built as part of the Java build.  It is reused from the dependency that libcudf is using.

Note that the nvcomp 1.x API is no longer supported by the CUDF Java JNI, so this is a breaking change.  Spark-Rapids currently depends on the nvcomp 1.x JNI.  I will be putting up a PR for spark-rapids that changes it to use this 2.x JNI, which will need to be merged soon after this PR to minimize incompatibilities.

Authors:
  - Jim Brennan (https://github.com/jbrennan333)

Approvers:
  - Alessandro Bellina (https://github.com/abellina)
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/9384