Add NVTX Start/End Ranges to JNI (#9563)
This adds start/end ranges to cuDF JNI. These types of ranges can be started and stopped in different threads, and the creation and stopping does not need to follow the order that a push/pop range follows (when a start/end range is closed, the actual range is closed, not whoever was started last).

This adds a boolean and a long to `NvtxRange` as a class. It may be better to create a `NvtxRangeStartEnd`, and even expose some helper factory-like methods in `NvtxRange`. Looking for ideas on whether it is preferred as implemented, or if we want to break it off into a different instance.

Authors:
  - Alessandro Bellina (https://github.com/abellina)

Approvers:
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/9563